+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog of ai-cockpit.
 
 ### Features
 
+-  action automation and decision detail shortcuts ([f3b88](https://github.com/starwit/ai-cockpit/commit/f3b88403227c0fb) Anett Hübner)  
 -  ab#1378: Added autonomy level information ([b22a0](https://github.com/starwit/ai-cockpit/commit/b22a0dfd25a1962) Anett Hübner)  
 -  added dialog for automomy level ([22660](https://github.com/starwit/ai-cockpit/commit/2266050b8efd97f) Anett Hübner)